Refactored the changes
