+++ support-e2e/playwright.local.config.ts
@@ -3,7 +3,8 @@ import { baseObject } from "./playwright.base.config";
 
 export default defineConfig({
   ...baseObject,
-  fullyParallel: true,//To prevent tests from running parallel while testing locally as it causes timeouts
+  fullyParallel: true,//To prevent tests from running parallel while testing locally change this to false
+  retries: process.env.CI ? 2 : 2,
   use: {
     /* Maximum time each action such as `click()` can take. Defaults to 0 (no limit). */
     actionTimeout: 0,
