+++ support-e2e/tests/recurringContributions.test.ts
@@ -1,8 +1,14 @@
 import { expect, test } from "@playwright/test";
 import { setTestCookies } from "./utils/cookies";
-import { email, firstName, lastName } from "./utils/users";
+import {  firstName  } from "./utils/users";
+import {setTestUserDetails} from "./utils/testUserDetails";
+import {checkRecaptcha} from "./utils/recaptcha";
+import {fillInCardDetails} from "./utils/cardDetails";
+import {fillInDirectDebitDetails} from "./utils/directDebitDetails";
+
 
 test.beforeEach(async ({ page, context, baseURL }) => {
+
   const baseUrlWithFallback = baseURL ?? "https://support.theguardian.com";
   // We should remove the forcing into the ab test once this has been made live
   const pageUrl = `${baseUrlWithFallback}/uk/contribute#ab-twoStepCheckoutWithNudgeBelow=variant_a`;
@@ -14,46 +20,17 @@ test.beforeEach(async ({ page, context, baseURL }) => {
 
 test.describe("Sign up for a Recurring Contribution (New Contributions Flow)", () => {
   test("Monthly contribution sign-up with Stripe - GBP", async ({ page }) => {
+
     await page.getByRole("tab").getByText("Monthly").click();
     await page.getByRole("button", { name: "Continue to checkout" }).click();
 
-    await page.getByLabel("Email address").type(email);
-    await page.getByLabel("First name").type(firstName);
-    await page.getByLabel("Last name").type(lastName);
+    await setTestUserDetails(page);
 
     await page.getByLabel("Credit/Debit card").click();
 
-    // it would be nice to use aria style selectors here, but given Stripes
-    // very secure implementation, it is quite hard
-    await page
-      .frameLocator("#cardNumber iframe")
-      .locator('input[name="cardnumber"]')
-      .fill("4242424242424242");
+    await fillInCardDetails(page);
 
-    await page
-      .frameLocator("#expiry iframe")
-      .locator('input[name="exp-date"]')
-      .fill("01/50");
-
-    await page
-      .frameLocator("#cvc iframe")
-      .locator('input[name="cvc"]')
-      .fill("123");
-
-    await expect(
-      await page
-        .frameLocator('[title="reCAPTCHA"]')
-        .locator("#recaptcha-anchor-label"),
-    ).toBeVisible();
-
-    const recaptchaIframe = page.frameLocator('[title="reCAPTCHA"]');
-    const recaptchaCheckbox = recaptchaIframe.locator(".recaptcha-checkbox");
-    await expect(recaptchaCheckbox).toBeEnabled();
-
-    await recaptchaCheckbox.click();
-    await expect(
-      recaptchaIframe.locator("#recaptcha-accessible-status"),
-    ).toContainText("You are verified");
+    await checkRecaptcha(page);
 
     await page.getByText(/Pay £([0-9]+) per month/).click();
 
@@ -66,33 +43,13 @@ test.describe("Sign up for a Recurring Contribution (New Contributions Flow)", (
     await page.getByRole("tab").getByText("Monthly").click();
     await page.getByRole("button", { name: "Continue to checkout" }).click();
 
-    await page.getByLabel("Email address").type(email);
-    await page.getByLabel("First name").type(firstName);
-    await page.getByLabel("Last name").type(lastName);
+    await setTestUserDetails(page);
 
     await page.getByRole('radio', { name: 'Direct debit' }).click();
 
-    await page.getByLabel("Account name").type("CP Scott");
-    await page.getByLabel("Sort code").type("200000");
-    await page.getByLabel("Account number").type("55779911");
+    await fillInDirectDebitDetails(page);
 
-    await page.getByRole('checkbox', { name: 'I confirm that I am the account holder and I am solely able to\n' +
-        '\t\t\t\t\t\t\tauthorise debit from the account' }).check();
-
-    await expect(
-      await page
-        .frameLocator('[title="reCAPTCHA"]')
-        .locator("#recaptcha-anchor-label"),
-    ).toBeVisible();
-
-    const recaptchaIframe = page.frameLocator('[title="reCAPTCHA"]');
-    const recaptchaCheckbox = recaptchaIframe.locator(".recaptcha-checkbox");
-    await expect(recaptchaCheckbox).toBeEnabled();
-
-    await recaptchaCheckbox.click();
-    await expect(
-      recaptchaIframe.locator("#recaptcha-accessible-status"),
-    ).toContainText("You are verified");
+    await checkRecaptcha(page);
 
     await page.getByText(/Pay £([0-9]+) per month/).click();
 
@@ -103,6 +60,7 @@ test.describe("Sign up for a Recurring Contribution (New Contributions Flow)", (
 
 test.describe("Sign up for a Recurring Contribution (New Contributions Flow)", () => {
   test("Annual contribution sign-up with Stripe - USD", async ({ page,baseURL,context }) => {
+
     const baseUrlWithFallback = baseURL ?? "https://support.theguardian.com";
     const pageUrl = `${baseUrlWithFallback}/us/contribute#ab-twoStepCheckoutWithNudgeBelow=variant_a`;
 
@@ -110,49 +68,22 @@ test.describe("Sign up for a Recurring Contribution (New Contributions Flow)", (
     await setTestCookies(context, firstName, domain);
     await page.goto(pageUrl, { waitUntil: "networkidle" });
 
+    // await page.getByRole("button", { name: "Select a country" }).click();
+    // await page.getByRole('link', { name: 'United States US$' }).click();
+
     await page.getByRole("tab").getByText("Annual").click();
     await page.getByRole("button", { name: "Continue to checkout" }).click();
 
-    await page.getByLabel("Email address").type(email);
-    await page.getByLabel("First name").type(firstName);
-    await page.getByLabel("Last name").type(lastName);
+    await setTestUserDetails(page);
 
     await page.getByLabel("State").type("NY");
     await page.getByLabel("ZIP code").type("90210");
 
     await page.getByRole('radio', { name: 'Credit/Debit card' }).click();
 
-    // it would be nice to use aria style selectors here, but given Stripes
-    // very secure implementation, it is quite hard
-    await page
-      .frameLocator("#cardNumber iframe")
-      .locator('input[name="cardnumber"]')
-      .fill("4242424242424242");
+    await fillInCardDetails(page);
 
-    await page
-      .frameLocator("#expiry iframe")
-      .locator('input[name="exp-date"]')
-      .fill("0150");
-
-    await page
-      .frameLocator("#cvc iframe")
-      .locator('input[name="cvc"]')
-      .fill("111");
-
-    await expect(
-      await page
-        .frameLocator('[title="reCAPTCHA"]')
-        .locator("#recaptcha-anchor-label"),
-    ).toBeVisible();
-
-    const recaptchaIframe = page.frameLocator('[title="reCAPTCHA"]');
-    const recaptchaCheckbox = recaptchaIframe.locator(".recaptcha-checkbox");
-    await expect(recaptchaCheckbox).toBeEnabled();
-
-    await recaptchaCheckbox.click();
-    await expect(
-      recaptchaIframe.locator("#recaptcha-accessible-status"),
-    ).toContainText("You are verified");
+    await checkRecaptcha(page);
 
     await page.locator('button:has-text("Support us with")').click();
