+++ support-e2e/tests/utils/cardDetails.ts
@@ -1,0 +1,22 @@
+import { Page } from "@playwright/test";
+
+export const fillInCardDetails = async (
+  page:Page,
+) => {
+  // it would be nice to use aria style selectors here, but given Stripes
+  // very secure implementation, it is quite hard
+  await page
+    .frameLocator("#cardNumber iframe")
+    .locator('input[name="cardnumber"]')
+    .fill("4242424242424242");
+
+  await page
+    .frameLocator("#expiry iframe")
+    .locator('input[name="exp-date"]')
+    .fill("01/50");
+
+  await page
+    .frameLocator("#cvc iframe")
+    .locator('input[name="cvc"]')
+    .fill("123");
+};
